Added groups, former groups and logevents to CoordinatorAdmin
Split admin.py to separate modules

Closes #296 #297
