+++ api/models/group.py
@@ -119,7 +119,7 @@ class Group(GroupCommon):
         teacher_names = ",".join(t.personal_info.full_name for t in self.teachers.all())
         return (
             f"Group {self.pk}, {self.language_and_level} (coordinators: {coordinator_names}, "
-            f"teachers: {teacher_names}, {self.students.count()} students."
+            f"teachers: {teacher_names}, {self.students.count()} students.)"
         )
 
     def _teachers_group_count_annotation(self) -> QuerySet[Teacher]:
